rework initialization code
swfdec_player_set_loader() no longer calls into the script engine but requires
a call to swfdec_player_advance() now.
swfdec_player_advance in turn was changed to allow triggering timeouts
immediately. So if something external happens, it can happen _now_ and not
_almost now_, as it was before.
There's some unfortunate side effects:
1) swfdec_player_get_next_event() can now return 0 which was reserved for
   "nothing is happening" before. TO cope with that, it returns a signed long
   now where a return value of -1 indicates this.
2) It is no longer possible to detect if the file can be played by Swfdec
   directly after calling swfdec_player_set_loader(), since the player will
   never be initialized by then.

+++ libswfdec/swfdec_player.h
@@ -77,7 +77,7 @@ void		swfdec_player_set_loader_with_variables
 						 const char *	variables);
 
 gboolean	swfdec_player_is_initialized	(SwfdecPlayer *	player);
-guint		swfdec_player_get_next_event  	(SwfdecPlayer *	player);
+glong		swfdec_player_get_next_event  	(SwfdecPlayer *	player);
 double		swfdec_player_get_rate		(SwfdecPlayer *	player);
 void		swfdec_player_get_image_size	(SwfdecPlayer *	player,
 						 int *		width,
@@ -107,7 +107,7 @@ void		swfdec_player_render		(SwfdecPlayer *	player,
 						 double		width,
 						 double		height);
 void		swfdec_player_advance		(SwfdecPlayer *	player,
-						 guint		msecs);
+						 gulong		msecs);
 gboolean	swfdec_player_handle_mouse	(SwfdecPlayer *	player, 
 						 double		x,
 						 double		y,
